warrior-support-system/python-backend/requirements.txt
# Core FastAPI and web framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6


# Data processing and analysis
pandas==2.1.3
numpy==1.25.2
joblib==1.3.2

# Audio processing
soundfile==0.12.1
openai-whisper==20231117
librosa==0.10.1

# Computer vision
opencv-python==4.8.1.78
Pillow==10.1.0

# Plotting and visualization
plotly==5.17.0
matplotlib==3.8.2
seaborn==0.13.0

# Machine learning
scikit-learn==1.3.2
torch==2.1.1
torchvision==0.16.1
torchaudio==2.1.1
transformers==4.35.2

# Natural language processing
nltk==3.8.1
spacy==3.7.2
textblob==0.17.1

# Database
sqlalchemy==2.0.23

# Utilities
python-dotenv==1.0.0
requests==2.31.0
aiofiles==23.2.0
pydantic==2.5.0

# Development
pytest==7.4.3
black==23.11.0
flake8==6.1.0

streamlit
